refactor: MVC에 따라 폴더 재구조화, 클래스 명칭 수정
